## Describe your changes
basic high priority for post method back end complete, refactored to fix
bug with post function with Linda's original refactoring.
4 passed test base functions, 1 warning 3 failed for lower priority task
that needs logic to be confirm with rest of team.

implementation for merge feature working however will need to reconfirm
with John in regards to test cases to refine in next PR

## Issue ticket number and link
https://fireapp-emergiq-2024.atlassian.net/browse/FE-151

+++ domain/entity/question.py
@@ -1,6 +1,5 @@
 from datetime import datetime
 from sqlalchemy import Column, Integer, String, DateTime, Boolean, Enum, Text
-from sqlalchemy.dialects.mysql import TINYINT
 from domain.base import Base
 from domain.type import QuestionType
 
@@ -14,7 +13,7 @@ class Question(Base):
     role = Column(String(256), nullable=True)
     description = Column(String(1024), nullable=False)
     choice = Column(Text, nullable=False)
-    difficulty = Column(TINYINT, default=1, nullable=False)
+    difficulty = Column(Integer, default=1, nullable=False)
     answer = Column(String(20), nullable=False)
     status = Column(Boolean, default=True, nullable=False)
     answered_time = Column(Integer, default=0, nullable=False)
